requirements.txt
asgiref==3.4.1
Django==3.2.13
django-extensions==3.1.5
django-model-utils==4.2.0
plotly==5.8.0
pytz==2022.1
sqlparse==0.4.2
tenacity==8.0.1
typing-extensions==4.1.1
